[Refactor]#2 presentation package 생성. 패키지 정리
- Clean Architecture로 리팩토링함에 따라 presentation 패키지 생성 후 파일 이동
