+++ app/src/main/AndroidManifest.xml
@@ -6,7 +6,7 @@
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 
     <application
-        android:name=".util.AppHelper"
+        android:name=".presentation.application.KakaoGalleryApplication"
         android:allowBackup="true"
         android:icon="@mipmap/app_icon"
         android:label="@string/app_name"
@@ -16,7 +16,7 @@
         <activity
             android:exported="true"
             android:theme="@style/SplashTheme"
-            android:name=".view.SplashActivity">
+            android:name=".presentation.ui.splash.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -25,7 +25,7 @@
         </activity>
 
         <activity
-            android:name=".view.MainActivity"
+            android:name=".presentation.ui.main.MainActivity"
             android:exported="false"/>
     </application>
