+++ app/src/main/java/com/gallery/kakaogallery/BaseRepository.kt
@@ -1,7 +1,7 @@
 package com.gallery.kakaogallery
 
-import com.gallery.kakaogallery.util.AppHelper
+import com.gallery.kakaogallery.presentation.application.KakaoGalleryApplication
 
 abstract class BaseRepository {
-    protected val TAG = AppHelper.getTag(this::class.java)
+    protected val TAG = KakaoGalleryApplication.getTag(this::class.java)
 }
