+++ app/src/main/java/com/gallery/kakaogallery/presentation/ui/main/FragmentHandler.kt
@@ -1,0 +1,7 @@
+package com.gallery.kakaogallery.presentation.ui.main
+
+import com.gallery.kakaogallery.presentation.ui.comp.HeaderComp
+
+interface FragmentHandler {
+    fun getHeaderCompForChange() : HeaderComp?
+}
