+++ app/src/main/java/com/gallery/kakaogallery/presentation/ui/main/MainViewModel.kt
@@ -1,0 +1,8 @@
+package com.gallery.kakaogallery.presentation.ui.main
+
+import com.gallery.kakaogallery.presentation.ui.base.BaseViewModel
+
+class MainViewModel : BaseViewModel() {
+
+    var currentPage : Int = 0
+}
