+++ app/src/main/java/com/gallery/kakaogallery/util/BaseUtil.kt
@@ -1,5 +1,7 @@
 package com.gallery.kakaogallery.util
 
+import com.gallery.kakaogallery.presentation.application.KakaoGalleryApplication
+
 abstract class BaseUtil {
-    protected val TAG = AppHelper.getTag(this::class.java)
+    protected val TAG = KakaoGalleryApplication.getTag(this::class.java)
 }
