--- app/src/main/java/com/gallery/kakaogallery/view/MainViewModel.kt
@@ -1,8 +1,0 @@
-package com.gallery.kakaogallery.view
-
-import com.gallery.kakaogallery.view.base.BaseViewModel
-
-class MainViewModel : BaseViewModel() {
-
-    var currentPage : Int = 0
-}
