--- app/src/main/java/com/gallery/kakaogallery/view/SplashActivity.kt
@@ -1,24 +1,0 @@
-package com.gallery.kakaogallery.view
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
-
-class SplashActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        Handler(Looper.getMainLooper()).postDelayed({
-            goNextView()
-        }, 1000)
-    }
-
-    private fun goNextView(){
-        val intent = Intent(this, MainActivity::class.java).apply {
-            addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK)
-            addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-        }
-        startActivity(intent)
-    }
-}
