--- app/src/main/java/com/gallery/kakaogallery/view/base/FragmentHandler.kt
@@ -1,7 +1,0 @@
-package com.gallery.kakaogallery.view.base
-
-import com.gallery.kakaogallery.view.comp.HeaderComp
-
-interface FragmentHandler {
-    fun getHeaderCompForChange() : HeaderComp?
-}
