+++ app/src/main/res/layout/fragment_search_image.xml
@@ -4,7 +4,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.gallery.kakaogallery.view.searchimages.SearchImageViewModel" />
+            type="com.gallery.kakaogallery.presentation.ui.searchimage.SearchImageViewModel" />
     </data>
 
     <RelativeLayout
